Refactor CategoryAchievement and added icons
