+++ Source/BusinessLogic/Logic/Achievements/WargamerAchievement.cs
@@ -1,12 +1,10 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using BusinessLogic.DataAccess;
-using BusinessLogic.Models;
 using BusinessLogic.Models.Achievements;
 
 namespace BusinessLogic.Logic.Achievements
 {
-    public class WargamerAchievement : BaseAchievement
+    public class WargamerAchievement : CategoryAchievement
     {
         public WargamerAchievement(IDataContext dataContext) : base(dataContext)
         {
@@ -29,32 +27,6 @@ namespace BusinessLogic.Logic.Achievements
             {AchievementLevel.Gold, 30}
         };
 
-        public override AchievementAwarded IsAwardedForThisPlayer(int playerId)
-        {
-            var result = new AchievementAwarded
-            {
-                AchievementId = Id
-            };
-
-            var allDiceGames =
-                DataContext
-                    .GetQueryable<PlayerGameResult>()
-                    .Where(x => x.PlayerId == playerId
-                        && x.PlayedGame.GameDefinition.BoardGameGeekGameDefinition.Categories.Any(o => o.CategoryName == "Wargame"))
-                    .Select(x => x.PlayedGame.GameDefinitionId)
-                    .Distinct()
-                    .ToList();
-
-            result.PlayerProgress = allDiceGames.Count;
-            result.RelatedEntities = allDiceGames.ToList();
-
-            if (result.PlayerProgress < LevelThresholds[AchievementLevel.Bronze])
-            {
-                return result;
-            }
-
-            result.LevelAwarded = GetLevelAwarded(result.PlayerProgress);
-            return result;
-        }
+        public override string CategoryName => "Wargame";
     }
 }
