+++ Source/UI/sass/imports/fonts.scss
@@ -89,3 +89,23 @@ $custom-font-path: $fa-font-path;
 .ns-icon-dice-chucker:before {
     content: "\e900";
 }
+
+.ns-icon-civbuilder:before {
+    content: "\e911";
+}
+
+.ns-icon-brains:before {
+    content: "\e90f";
+}
+
+.ns-icon-mastershuffler:before {
+    content: "\e910";
+}
+
+.ns-icon-categorized:before {
+    content: "\e90e";
+}
+
+.ns-icon-wargamer:before {
+    content: "\e90d";
+}
